Refactor Localizers and RobotConfig; remove docs
Rename and reorganize localization code and update odometry constants. Packages changed from core.localization to core.Localizers; localizer classes (Two/ThreeDeadWheel, Pinpoint, MotorEncoder, Dev) updated imports and RobotConfig references. CrawlerLocaliser API: rename resetPoser to resetPose and add resetPose(Pose2d) overload; DevLocaliser updated accordingly. RobotConfig moved odometry-related constants into RobotConfig.Odometry, updated usages and added TICKS_PER_CM. CrawlerRobot imports adjusted and ThreeDeadWheelLocaliser used for dead-wheel localization. Also remove many docs/wiki/media/legal files from the repository.
