+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Crawler/core/Localizers/CrawlerLocaliser.java
@@ -5,6 +5,7 @@ import com.arcrobotics.ftclib.geometry.Pose2d;
 public interface CrawlerLocaliser {
     void update();
     Pose2d getPose();
-    void resetPoser();
+    void resetPose();
 
+    void resetPose(Pose2d pose);
 }
