+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Crawler/core/Localizers/DevLocaliser.java
@@ -43,8 +43,10 @@ public class DevLocaliser  implements CrawlerLocaliser{
     }
 
     @Override
-    public void resetPoser() {
+    public void resetPose() {
         pose = new Pose2d(0, 0, new Rotation2d(0));
         lastTime = System.nanoTime();
     }
+
+
 }
