--- docs/README.md
@@ -1,44 +1,0 @@
-# FTC Robot Controller - GitBook Documentation
-
-Welcome to the FTC Robot Controller GitBook documentation! This comprehensive guide covers everything you need to know about robotics, control systems, and optimization.
-
-## Table of Contents
-
-1. **Introduction** - Overview of the FTC control system
-2. **Getting Started** - Initial setup and configuration
-3. **Robot Tuning** - Hardware optimization and motor/servo calibration
-4. **General Tuning** - Software parameters and control adjustments
-5. **Troubleshooting** - Common problems and solutions
-6. **Advanced Topics** - Performance optimization and custom configurations
-
-## What's Inside?
-
-### For Team Leads & Mentors
-- System architecture overview
-- Hardware configuration guidelines
-- Performance benchmarking
-
-### For Programmers
-- Control system basics
-- Motor control and PID tuning
-- Navigation and autonomous programming
-
-### For Hardware Teams
-- Motor selection and calibration
-- Servo configuration
-- System diagnostics
-
-## Getting Start Fast
-
-Start with the **[Introduction](introduction.md)** section to understand the basics, then proceed to the guide most relevant to your needs.
-
-## Additional Resources
-
-- [Official FIRST Tech Challenge Docs](https://ftc-docs.firstinspires.org/)
-- [FTC Community Forum](https://www.firstinspires.org/community/technical-support)
-- [GitHub Repository](https://github.com/FIRST-Tech-Challenge/FtcRobotController)
-
----
-
-**Version**: 2025-2026 Season  
-**Last Updated**: February 2026
