--- docs/SUMMARY.md
@@ -1,25 +1,0 @@
-# Documentation Summary
-
-## Getting Started
-
-* [Introduction](introduction.md)
-
-## Guides
-
-* [Robot Tuning Guide](guides/robot-tuning-guide.md)
-  * Motor Tuning
-  * Servo Tuning
-  * Sensor Verification
-  * Power Management
-  
-* [General Tuning Guide](guides/general-tuning-guide.md)
-  * Motor Control Parameters
-  * Navigation Tuning
-  * Sensor-Based Tuning
-  * Autonomous Routine Tuning
-
-## Resources
-
-* [Official FTC Documentation](https://ftc-docs.firstinspires.org/)
-* [GitHub Repository](https://github.com/FIRST-Tech-Challenge/FtcRobotController)
-* [FIRST Community Support](https://www.firstinspires.org/community/technical-support)
