--- docs/book.json
@@ -1,20 +1,0 @@
-{
-  "title": "FTC Robot Controller Documentation",
-  "description": "Comprehensive guide for FTC robot programming and tuning",
-  "author": "FIRST Tech Challenge Community",
-  "gitbook": ">=3.0.0",
-  "root": "./",
-  "plugins": ["github", "anchors"],
-  "pluginsConfig": {
-    "github": {
-      "url": "https://github.com/FIRST-Tech-Challenge/FtcRobotController"
-    }
-  },
-  "structure": {
-    "readme": "README.md",
-    "summary": "SUMMARY.md"
-  },
-  "variables": {
-    "ftc_version": "2025-2026"
-  }
-}
