--- wiki/General-Tuning-Guide.md
@@ -1,36 +1,0 @@
-# General Tuning Guide
-
-*For detailed information on software optimization, see [General Tuning Guide in GitBook](../docs/guides/general-tuning-guide.md)*
-
-This is a quick reference for software tuning. The comprehensive guide covers:
-
-- Power scaling and deadzone
-- Motor acceleration control
-- Navigation tuning (turning, speed profiles)
-- Sensor response time calibration
-- Autonomous routine timing
-- Error recovery strategies
-
-## Quick Start
-
-1. **Configure Deadzone** - Adjust minimum input to prevent drift
-2. **Set Power Scaling** - Ensure responsive gamepad control
-3. **Calibrate Movement Timing** - Match programmed duration to actual movement
-4. **Tune Turning Speed** - Optimize rotation responsiveness
-5. **Verify Sensor Thresholds** - Test autonomous triggers in real conditions
-
-## Common Control Issues
-
-- **Robot drifts with stick centered**: Increase deadzone
-- **Sluggish response to input**: Increase power scaling
-- **Jerky acceleration**: Adjust acceleration ramp timing
-- **Autonomous overshoots**: Reduce motor power or add braking
-
-## Performance Optimization Tips
-
-- Start conservative; increase speeds gradually
-- Change one parameter per test session
-- Use telemetry to monitor real values
-- Test on the actual competition surface when possible
-
-See the complete [General Tuning Guide](../docs/guides/general-tuning-guide.md) for detailed procedures and examples.
