--- wiki/Home.md
@@ -1,30 +1,0 @@
-# FTC Robot Controller - GitHub Wiki
-
-Welcome to the FTC Robot Controller Wiki! This is your comprehensive guide to understanding, setting up, and optimizing your FTC robot control system.
-
-## Quick Navigation
-
-- **[Getting Started](Getting-Started)** - Set up your development environment and understand the basics
-- **[Robot Tuning Guide](Robot-Tuning-Guide)** - Optimize hardware behavior and motor/servo performance
-- **[General Tuning Guide](General-Tuning-Guide)** - Fine-tune software settings and control parameters
-- **[Troubleshooting](Troubleshooting)** - Common issues and solutions
-- **[Documentation](https://ftc-docs.firstinspires.org/)** - Official FIRST Tech Challenge Documentation
-
-## What is This?
-
-The FTC Robot Controller is the Android application that runs on your robot's control hub. This wiki provides practical guides to help you:
-
-- Understand how your robot system works
-- Configure motors and servos properly
-- Adjust PID controllers and movement parameters
-- Get the best performance from your hardware
-
-## For First-Time Users
-
-Start with our [Getting Started](Getting-Started) guide if you're new to FTC programming or robot control systems.
-
-## Need More Help?
-
-- Check the [Official FTC Documentation](https://ftc-docs.firstinspires.org/)
-- Visit the [FIRST Tech Challenge Community](https://www.firstinspires.org/community/technical-support)
-- Review the project [README](../README.md) for setup instructions
