--- wiki/_Sidebar.md
@@ -1,27 +1,0 @@
-## FTC Robot Controller Wiki
-
-### Getting Started
-- [Home](Home)
-- [Getting Started Guide](Getting-Started)
-
-### Guides
-- [Robot Tuning Guide](Robot-Tuning-Guide)
-  - Motor Tuning
-  - Servo Tuning
-  - Sensor Verification
-  - Power Management
-  
-- [General Tuning Guide](General-Tuning-Guide)
-  - Motor Control Parameters
-  - Navigation Tuning  
-  - Sensor-Based Tuning
-  - Autonomous Tuning
-
-### Support
-- [Troubleshooting Guide](Troubleshooting)
-- [Official FTC Docs](https://ftc-docs.firstinspires.org/)
-
----
-
-**Version**: 2025-2026 Season  
-**Last Updated**: February 2026
